requirements.txt
Package    Version
---------- -------
emoji      2.2.0
numpy      1.24.3
Pillow     9.5.0
pip        22.3.1
setuptools 65.5.0
sv-ttk     2.4.5